# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Credd AI - Financial infrastructure for autonomous AI agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/agents/{agentId}/transactions:
    get:
      operationId: listAgentTransactions
      tags: [transactions]
      summary: List transactions for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /v1/agents/{agentId}/policies:
    get:
      operationId: listAgentPolicies
      tags: [policies]
      summary: List policies for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
...
  /v1/accounts/{accountId}/transactions:
    get:
      operationId: listAccountTransactions
      tags: [transactions]
      summary: List transactions for an account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /v1/accounts/{accountId}/policies:
    get:
      operationId: listAccountPolicies
      tags: [policies]
      summary: List policies for an account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
...
  /v1/cards/{cardId}/policies:
    get:
      operationId: listCardPolicies
      tags: [policies]
      summary: List policies for a card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
  /v1/policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: List all policies
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Create a policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyBody"
      responses:
        "201":
          description: Created policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyResponse"
  /v1/policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: Get policy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyResponse"
  /v1/policies/{id}/status:
    patch:
      operationId: updatePolicyStatus
      tags: [policies]
      summary: Update policy status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyResponse"
  /v1/policies/{id}/assignments:
    get:
      operationId: listPolicyAssignments
      tags: [policies]
      summary: List assignments for a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyAssignmentListResponse"
...
  /v1/transactions/attempt:
    post:
      operationId: attemptTransaction
      tags: [transactions]
      summary: Attempt a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptTransactionBody"
      responses:
        "200":
          description: Transaction attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionAttemptResponse"
  /v1/transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
  /v1/transactions/{id}/approval:
    get:
      operationId: getTransactionApproval
      tags: [approvals]
      summary: Get approval for a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approval request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
...
components:
  schemas:
...
    Policy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [SPEND_LIMIT, MERCHANT_ALLOWLIST, MERCHANT_BLOCKLIST, APPROVAL_REQUIRED, TIME_WINDOW]
        status:
          type: string
          enum: [active, disabled]
        configJson:
          type: object
        description:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - status
        - configJson
        - createdAt
        - updatedAt
    PolicyAssignment:
      type: object
      properties:
        id:
          type: string
        policyId:
          type: string
        targetType:
          type: string
          enum: [AGENT, ACCOUNT, CARD]
        targetId:
          type: string
        priority:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - policyId
        - targetType
        - targetId
        - priority
        - createdAt
    Transaction:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        accountId:
          type: string
        cardId:
          type: string
          nullable: true
        merchantId:
          type: string
          nullable: true
        direction:
          type: string
          enum: [DEBIT, CREDIT]
        amountMinor:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [attempted, pending, approved, declined, settled, reversed]
        decision:
          type: string
          enum: [ALLOW, DENY, REVIEW, NONE]
        approvalStatus:
          type: string
          enum: [NOT_REQUIRED, PENDING, APPROVED, DECLINED]
        declinedReason:
          type: string
          nullable: true
        requestedAt:
          type: string
        processedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - agentId
        - accountId
        - direction
        - amountMinor
        - currency
        - status
        - decision
        - approvalStatus
        - requestedAt
        - createdAt
        - updatedAt
    ApprovalRequest:
      type: object
      properties:
        id:
          type: string
        transactionId:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, DECLINED, EXPIRED]
        requestedByType:
          type: string
        requestedReason:
          type: string
          nullable: true
        reviewerId:
          type: string
          nullable: true
        reviewedAt:
          type: string
          nullable: true
        decisionReason:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - transactionId
        - status
        - requestedByType
        - createdAt
        - updatedAt
...
    MatchedPolicy:
      type: object
      properties:
        policyId:
          type: string
        policyType:
          type: string
        outcome:
          type: string
          enum: [ALLOW, DENY, REVIEW]
      required:
        - policyId
        - policyType
        - outcome
    TransactionDecisionResult:
      type: object
      properties:
        result:
          type: string
          enum: [ALLOW, DENY, REVIEW]
        reason:
          type: string
        matchedPolicies:
          type: array
          items:
            $ref: "#/components/schemas/MatchedPolicy"
      required:
        - result
        - reason
        - matchedPolicies

    # Request bodies
...
    CreatePolicyBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [SPEND_LIMIT, MERCHANT_ALLOWLIST, MERCHANT_BLOCKLIST, APPROVAL_REQUIRED, TIME_WINDOW]
        configJson:
          type: object
        description:
          type: string
      required:
        - name
        - type
        - configJson
...
    AttemptTransactionBody:
      type: object
      properties:
        agentId:
          type: string
        accountId:
          type: string
        cardId:
          type: string
        merchantId:
          type: string
        amountMinor:
          type: integer
        currency:
          type: string
        direction:
          type: string
          enum: [DEBIT, CREDIT]
      required:
        - agentId
        - accountId
        - amountMinor
        - currency
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    PolicyResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Policy"
      required:
        - success
        - data
    PolicyListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Policy"
      required:
        - success
        - data
...
    PolicyAssignmentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/PolicyAssignment"
      required:
        - success
        - data
    TransactionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Transaction"
      required:
        - success
        - data
    TransactionListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required:
        - success
        - data
    TransactionAttemptResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            transaction:
              $ref: "#/components/schemas/Transaction"
            decision:
              $ref: "#/components/schemas/TransactionDecisionResult"
            approvalRequest:
              $ref: "#/components/schemas/ApprovalRequest"
              nullable: true
          required:
            - transaction
            - decision
      required:
        - success
        - data
    ApprovalResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ApprovalRequest"
      required:
        - success
        - data
...